// ticktacktoe-ui/src/app/types/round-result.ts
export enum RoundResult {
    NotOver = "NotOver",
    Draw = "Draw",
    TopRow = "TopRow",
    MiddleRow = "MiddleRow",
    BottomROw = "BottomRow",
    LeftColumn = "LeftColumn",
    MiddleColumn = "MiddleColumn",
    RightColumn = "RightColumn",
    Diagonal = "Diagonal",
    InverseDiagonal = "InverseDiagonal"
}